[test|vectorarray] move indices using tests to strategies
